fixed client when the Raven URL is null
In the protocol, it is specified that the Raven URL can be null to disable the client. But that does not work well as the send() method relies on getting an array line 419.

To avoid a PHP notice, I've added a check to just return if the servers array is not configured.