$errcontext must not be modified.
The contents of the errcontext param of the PHP error handler must not
be modified since it may contain references. If control is returned to
the code which triggered the error (so a non-fatal error) we might
have changed important state.
This would be bad.

Raven_Stacktrace::get_stack_info can be passed this errcontext, and will
truncate strings in the array to a given limit, modifying errcontext in
the process. Whoops.