work around PHP bug #50688
if any descendant of Exception is instantiated from within a callback
passed into the various array sorting functions, then PHP will raise a
warning ("Array was modified by the user comparison function").

As Raven always registers with E_ALL and then checks whether it needs to
run in the handler itself, even if Raven decides not to do anything, it
still instantiates the exception and thus will still cause the warning.

For example, if you run with E_NOTICE off and thus register Raven to not
run on E_NOTICE, but then the user raises an E_NOTICE in a usort()
callback, php bug 50688 will cause a warning to be thrown (which then
likely *will* activate Raven on the second pass)