Add serialization for callables
If a value to be serialized for transport is of type
`callable`, serialize the return value instead of
resorting to string casting.

This is useful for lazy-loading values for transport.

Example:

A framework/application uses a hook-filter event system,
where the "current hook" changes constantly. By defining
a callable which returns the current hook ID, we can
define the Raven client context once and it will be
lazy-loaded upon serialization.

Otherwise we would have to hook into all possible hooks
via some magic and keep redefining the wanted context
value, while wasting processing power.

Possible pitfalls that need to be considered:

-   Recursive callables which create infinite loops
-   Callables that use too much system resources
-   Invalid parameters passed to callables (only allow
    callables that accept no parameters?)
-   Cases where the actual callable is wanted to be passed
    as a string instead of returning the result.