Allow extras to be cleared
I am trying to send along data (extras) that is specific to a particular error. Currently when I set tags it attaches them to ALL errors that I send. After I send an error I want to be able to clear out that data so it doesn't pollute future calls. This is easier than trying to build up a full new scope each time