Add PSR-3 logger to client builder and http transport
To be able to log Sentry PHP framework incidents,
a trait is introduced for Psr\Log\LoggerAwareInterface
which implements a protected getter, that falls back
to a NullLogger in case no logger has been set by the
code that created the object. This fallback is especially
important to have backwards compatibility.

ClientBuilder and HttpTransport now implement LoggerAwareInterface
by using the trait. Other classes can be adapted accordingly.