Remove final for State/Scope
Hi! I make error handler for different logic for different environments and other conditions.
While writing unit tests, I ran into the problem that I cannot make a mock for the State/Scope because it is final.
I propose to remove the final.

If you are against, then please add an interface that can be mocked.