Fatal Error Listener: Allow configuring of error limit
From the looks of the code, this was meant to be configurable, but
was never lifted up into a user option. We ran into a case where we needed
to bump the reserved memory (currently doing so with a custom integration).

This change does not affect the interface of registerOnceFatalErrorHandler
thus is backwards compatible.

I believe the code can be a bit cleaner if we can change the param to be ?int