Swap order of Retry and Error http client plugins
The ErrorPlugin is responsible for filtering out requests that returned http
error codes. In the original order this filtering would happen only after a
failing request to the sentry endpoint was retried some number of times, as
configured by `getSendAttempts`.

This meant that when sentry rate limits the requests from this client the
requests would be retried before resulting in a ClientErrorException.

This patch reorders the ErrorPlugin and the RetryPlugin so the
ClientErrorException is thrown before the sdk attempts to retry the request.

Fixes #1204