bug: Set sentry-trace header when using tracing middleware
According to PSR-7 (https://www.php-fig.org/psr/psr-7/), `withHeader` operates in the following manner:
```
Return an instance with the provided value replacing the specified header.
[...]
This method MUST be implemented in such a way as to retain the
immutability of the message, and MUST return an instance that has the
new and/or updated header and value.
```

Due to this, we must set `$request` to the modified request object (from the return value of
withHeader) in order to actually send the sentry-trace header.