Forbid serializing the HubAdapter in the first place
The symfony cache takes the hint when something isn't serializable. But it will fail if something successfully serializes and then fails to deserialize.

This small change makes the HubAdapter still work, when it accidentally get's caught in the Metadata Cache of Doctrine. This happens, for example, when you use a logger in doctrine events.