Scrub PII exposed by Cloudflare proxy services
CloudFlare's proxy services are known to inject 'CF-Connecting-IP' header in the request, which contains the real IP address of the requester. This header seems to have similar functionality as 'X-Forwarded-For' and/or 'X-Real-IP'. 

As with X-Real-IP and X-Forwarded-For, the IP-address contained in the header is considered PII and should be scrubbed by default. 

See also: https://developers.cloudflare.com/fundamentals/reference/http-headers/